backend/tests/requirements.txt
# Test requirements for Lumen backend testing

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.1
pytest-cov>=4.1.0

# HTTP testing
httpx>=0.24.1
fastapi[testing]>=0.100.0

# Database testing
pytest-postgresql>=5.0.0
sqlalchemy-utils>=0.41.1

# Mock and fixtures
responses>=0.23.1
faker>=19.0.0

# Coverage and reporting
coverage>=7.3.0
pytest-html>=3.2.0
pytest-xdist>=3.3.1

# Performance and profiling
pytest-benchmark>=4.0.0
memory-profiler>=0.61.0

# Security testing
bandit>=1.7.5
safety>=2.3.5

# Code quality
flake8>=6.0.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0